chore: move workspace functions to module
