+++ rap/src/bin/cargo-rap/cargo_check.rs
@@ -1,11 +1,10 @@
 use crate::args;
-use cargo_metadata::{
-    camino::{Utf8Path, Utf8PathBuf},
-    Metadata,
-};
+use cargo_metadata::camino::Utf8Path;
 use rap::utils::log::rap_error_and_exit;
-use std::{collections::BTreeMap, env, process::Command, time::Duration};
+use std::{env, process::Command, time::Duration};
 use wait_timeout::ChildExt;
+
+mod workspace;
 
 pub fn run() {
     match env::var("RAP_RECURSIVE")
@@ -13,9 +12,9 @@ pub fn run() {
         .map(|s| s.to_ascii_lowercase())
         .as_deref()
     {
-        Some("shallow") => shallow_run(),
-        Some("deep") => deep_run(),
         Some("none") | None => default_run(),
+        Some("deep") => workspace::deep_run(),
+        Some("shallow") => workspace::shallow_run(),
         _ => rap_error_and_exit(
             "`recursive` should only accept one the values: none, shallow or deep.",
         ),
@@ -71,96 +70,4 @@ fn cargo_check(dir: &Utf8Path) {
 /// Just like running a cargo check in a folder.
 fn default_run() {
     cargo_check(".".into());
-}
-
-/// Run cargo check in each member folder under current workspace.
-fn shallow_run() {
-    let cargo_toml = Utf8Path::new("Cargo.toml");
-    if !cargo_toml.exists() {
-        rap_error_and_exit("rap should be run in a folder directly containing Cargo.toml");
-    }
-    let ws_metadata = workspace(cargo_toml);
-    clean_and_check(&ws_metadata);
-}
-
-fn clean_and_check(ws_metadata: &Metadata) {
-    cargo_clean(&ws_metadata.workspace_root);
-    for pkg_folder in get_member_folders(ws_metadata) {
-        cargo_check(pkg_folder);
-    }
-}
-
-/// Usually run in workspace root before checking.
-fn cargo_clean(ws_root: &Utf8Path) {
-    rap_info!("cargo clean in workspace root {ws_root}");
-    if let Err(err) = Command::new("cargo")
-        .arg("clean")
-        .current_dir(ws_root)
-        .output()
-    {
-        rap_error_and_exit(format!("`cargo clean` exits unexpectedly:\n{err}"));
-    }
-}
-
-type Workspaces = BTreeMap<Utf8PathBuf, Metadata>;
-
-fn workspace(cargo_toml: &Utf8Path) -> Metadata {
-    let exec = cargo_metadata::MetadataCommand::new()
-        .manifest_path(cargo_toml)
-        .exec();
-    let metadata = match exec {
-        Ok(metadata) => metadata,
-        Err(err) => {
-            let err = format!(
-                "Failed to get the result of cargo metadata \
-                 in {cargo_toml}:\n{err}"
-            );
-            rap_error_and_exit(err)
-        }
-    };
-    metadata
-}
-
-fn get_member_folders(meta: &Metadata) -> Vec<&Utf8Path> {
-    meta.workspace_packages()
-        .iter()
-        .map(|pkg| pkg.manifest_path.parent().unwrap())
-        .collect()
-}
-
-/// Recursively run cargo check in each package folder from current folder.
-fn deep_run() {
-    let cargo_tomls = get_cargo_tomls_deep_recursively(".");
-    for ws_metadata in workspaces(&cargo_tomls).values() {
-        clean_and_check(ws_metadata);
-    }
-}
-
-fn get_cargo_tomls_deep_recursively(dir: &str) -> Vec<Utf8PathBuf> {
-    walkdir::WalkDir::new(dir)
-        .into_iter()
-        .filter_map(|entry| {
-            if let Ok(e) = entry {
-                if e.file_type().is_file() && e.file_name().to_str()? == "Cargo.toml" {
-                    let path = Utf8PathBuf::from_path_buf(e.into_path());
-                    return path.ok()?.canonicalize_utf8().ok();
-                }
-            }
-            None
-        })
-        .collect()
-}
-
-fn workspaces(cargo_tomls: &[Utf8PathBuf]) -> Workspaces {
-    let mut map = BTreeMap::new();
-    for cargo_toml in cargo_tomls {
-        let metadata = workspace(cargo_toml);
-        let root = &metadata.workspace_root;
-        // 每个 member package 解析的 workspace_root 和 members 是一样的
-        if !map.contains_key(root) {
-            map.insert(root.clone(), metadata);
-        }
-    }
-
-    map
 }
